Removed Cell class Added a better logging system Added navigation mouse support Fixed several issues with images not loading correctly if mouse was spammed right when application was started fixed deadlock when exiting application Download images only for previous and next X pages defined in a static variable. switched to warning level 4 and fixed jsoncpp warning TO DO: clear when exit the defined interval TO DO: integrate cmake TO DO: better animation when switching pages... maybe some ease in/out functionality

+++ ImageBrowser/HttpClient.h
@@ -44,7 +44,7 @@ private:
 	HttpClient();
 	~HttpClient();
 
-	static const int8_t NUM_THREADS;
+	static const size_t NUM_THREADS;
 	static const int64_t MAX_IMG_SIZE;
 
 	static HttpClient* mInstance;
